65c02: updates to filebrowser - fixed problems in filebrowser and cpmfs - updated Makefile for easier setup of machine - updated external-taliforth2 build script with a sanity check - updated external-cpm65 build script to reflect changes in CP/M 65 - updated included binaries from CP/M 65 - new hardware feature: RAM $0000-$03ff can be swapped pagewise - doc update to match changes

+++ src/tools/external-cpm65.sh
@@ -8,13 +8,26 @@ readonly BASEDIR="${PWD}"
 readonly BUILDTARGET="sorbus.zip"
 readonly TARGETDIR="${BASEDIR}/src/bin/cpm"
 
+for i in $(readlink -f ../llvm-mos) /usr /usr/local /opt/llvm-mos; do
+   if [ -x "${i}/bin/mos-cpm65-clang" ]; then
+      LLVM_PATH="${i}"
+      break
+   fi
+done
+
+if [ -z "${LLVM_PATH}" ]; then
+   echo "llvm-mos-sdk not found"
+   exit 1
+fi
+readonly LLVM_PATH
+
 set -e
 
 rm -rf "${WORKDIR}"
 mkdir -p "${WORKDIR}"
 cd "${WORKDIR}"
 git clone --depth 1 "${URL}" "${PWD}"
-make "${BUILDTARGET}"
+make "${BUILDTARGET}" LLVM="${LLVM_PATH}/bin"
 readonly ZIPDIR="${PWD}"
 cd "${TARGETDIR}"
 7z x -y "${ZIPDIR}/${BUILDTARGET}" CPMFS CPM BDOS
